requirements.txt
absl-py==0.8.1
aniso8601==9.0.1
asn1crypto==0.24.0
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
cached-property==1.5.2
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
click==8.0.4
colorama==0.4.3
cryptography==2.1.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
Flask-RESTful==0.3.9
gast==0.2.2
gevent==1.4.0
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==2.0.2
grpcio==1.25.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==2.0.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
keras-contrib==2.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
natsort==6.2.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandocfilters==1.4.2
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycrypto==2.6.1
Pygments==2.5.2
PyGObject==3.26.1
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2023.3
PyWavelets==1.1.1
pywinpty==0.5.7
pyxdg==0.25
PyYAML==5.2
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.3.2
SecretStorage==2.3.1
Send2Trash==1.5.0
six==1.16.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
zipp==0.6.0
